AI Financial Advisor/requirements.txt
streamlit==1.27.2
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
